src/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openai==2.8.0
